Don't use globals inside LRParser.
This causes problems in multithread parsing when "del" is executed, as
several error handlers might be in use at the same time. This is not a big
loss as errok and restart are readily available on parser instance, and
get_token is provided to ctor (and if not, can be easily known to code
instantiating).