Change write_table to atomically write to the parsetab file.
The previous implementation had issues with multiple python processes
simultaneously invoking `.yacc(...)` with the same arugments. One of them
could be midway through generating a .py file that the other process
would attempt to import and then crash with a `SyntaxError`. This diff
avoids that problem by instead writing to a tempfile and then atomically
renaming it. This means processes starting up at the same time will both
do the same work, but that's a lot better than having one of them crash!

`flanker` had this issue reported to them in
https://github.com/mailgun/flanker/issues/168 and they attempted to work
around this by committing the parsetab files that ply generates
(https://github.com/mailgun/flanker/pull/188), but this doesn't help if
the user is running a different version of ply than the version that
flanker generated their parsetab files with (because ply will go ahead
and regenerate those files).